refactor: AnswerHistory page
